requirements.txt
fastapi
uvicorn
pillow
beautifulsoup4
requests
openai
github-copilot-sdk
isodate
python-multipart
